Updated soft delete users to cover every user route closes #67 #33

+++ src/main/java/com/ga/showroom/service/UserService.java
@@ -78,6 +78,7 @@ public class UserService {
             userObject.setPassword(passwordEncoder.encode(userObject.getPassword()));
 
             userObject.setVerified(false);
+            userObject.setUserStatus(UserStatus.ACTIVE);
             userObject.setRole(Role.CUSTOMER); // Default user role
             User savedUser = userRepository.save(userObject);
 
@@ -110,18 +111,18 @@ public class UserService {
             myUserDetails = (MyUserDetails) authentication.getPrincipal();
             System.out.println("myUserDetails :::: "+myUserDetails.getUsername());
 
+            // Check if the user is inactive
+            if (myUserDetails.getUser().getUserStatus() == UserStatus.INACTIVE) {
+                return ResponseEntity
+                        .status(HttpStatus.FORBIDDEN)
+                        .body("Error: This account has been deactivated. Please contact an admin for support.");
+            }
+
             // Check if the user is verified
             if (!myUserDetails.getUser().getVerified()) { // assuming 'enabled' = email verified
                 return ResponseEntity
                         .status(HttpStatus.FORBIDDEN)
                         .body("Error: Email not verified. Please verify your email before logging in.");
-            }
-
-            // Check if the user is inactive
-            if (myUserDetails.getUser().getUserStatus() == UserStatus.INACTIVE) {
-                return ResponseEntity
-                        .status(HttpStatus.FORBIDDEN)
-                        .body("Error: This account has been deactivated.");
             }
 
             final String JWT = jwtUtils.generateJwtToken(myUserDetails);
@@ -146,6 +147,8 @@ public class UserService {
     public ChangePasswordResponse changePassword(ChangePasswordRequest changePasswordRequest) {
         User user = userRepository.findUserByEmailAddress(UserService.getCurrentLoggedInUser().getEmailAddress());
 
+        if (user.getUserStatus() == UserStatus.INACTIVE) throw new AccessDeniedException("This account has been deactivated. Please contact an admin for support.");
+
         if (!passwordEncoder.matches(changePasswordRequest.getOldPassword(), user.getPassword())) {
             return new ChangePasswordResponse("Old password incorrect");
         }
@@ -162,6 +165,8 @@ public class UserService {
     public UserProfile updateProfile(UserProfile userProfile, MultipartFile cprImage) {
         User user = userRepository.findUserByEmailAddress(UserService.getCurrentLoggedInUser().getEmailAddress());
 
+        if (user.getUserStatus().equals(UserStatus.INACTIVE)) throw new AccessDeniedException("This account has been deactivated. Please contact an admin for support.");
+
         if (getCurrentLoggedInUser().getRole().equals(Role.CUSTOMER)
                 && !user.getId().equals(getCurrentLoggedInUser().getId()))
             throw new AccessDeniedException("You are not authorized to change another user's profile data. Please contact a salesman or admin.");
@@ -174,6 +179,7 @@ public class UserService {
         profile.setHomeAddress(userProfile.getHomeAddress());
         profile.setCpr(userProfile.getCpr());
 
+        // TODO: Replace it with imageUpload() and imageDelete()
         // handle CPR image upload
         if (cprImage != null && !cprImage.isEmpty()) {
             try {
@@ -231,6 +237,9 @@ public class UserService {
         }
 
         User user = resetToken.getUser();
+
+        if (user.getUserStatus().equals(UserStatus.INACTIVE)) throw new AccessDeniedException("This account has been deactivated. Please contact an admin for support.");
+
         user.setPassword(passwordEncoder.encode(newPassword));
         userRepository.save(user);
 
@@ -271,8 +280,10 @@ public class UserService {
         }
 
         User user = verificationToken.getUser();
+
+        if (user.getUserStatus().equals(UserStatus.INACTIVE)) throw new AccessDeniedException("This account has been deactivated. Please contact an admin for support.");
+
         user.setVerified(true);
-        user.setUserStatus(UserStatus.ACTIVE);
 
         userRepository.save(user);
         emailVerificationTokenRepository.delete(verificationToken);
@@ -292,6 +303,8 @@ public class UserService {
 
         if (user == null) throw new InformationNotFoundException("User with email address " + emailAddress + " not found.");
 
+        if (user.getUserStatus().equals(UserStatus.INACTIVE)) throw new AccessDeniedException("This account has been deactivated. Please contact an admin for support.");
+
         user.setRole(role);
 
         return userRepository.save(user);
@@ -305,8 +318,30 @@ public class UserService {
                 .findById(userId)
                 .orElseThrow(() -> new InformationNotFoundException("User not found"));
 
+
+        if (user.getUserStatus().equals(UserStatus.INACTIVE)) throw new AccessDeniedException("This account has already been deactivated.");
+
         user.setUserStatus(UserStatus.INACTIVE);
         userRepository.save(user);
+    }
+
+    /**
+     * Reset user account status back to active.
+     * @param userId Long
+     * @return User
+     */
+    public User reactivateUserAccount(Long userId) {
+        if (!getCurrentLoggedInUser().getRole().equals(Role.ADMIN))
+            throw new AccessDeniedException("Only an admin is authorized to change user status.");
+
+        User user = userRepository
+                .findById(userId)
+                .orElseThrow(() -> new InformationNotFoundException("User with ID " + userId + " not found"));
+
+        if (user.getUserStatus().equals(UserStatus.ACTIVE)) throw new AccessDeniedException("This account is already activated.");
+
+        user.setUserStatus(UserStatus.ACTIVE);
+        return userRepository.save(user);
     }
 
 }
